Moved trajectory_visualization to rvizu_plugin_render_tools
