+++ visualization/rviz_plugin_render_tools/src/trajectory_visualization.cpp
@@ -34,7 +34,7 @@
 
 /* Author: Dave Coleman */
 
-#include <moveit/trajectory_rviz_plugin/trajectory_visualization.h>
+#include <moveit/rviz_plugin_render_tools/trajectory_visualization.h>
 
 #include <moveit/rviz_plugin_render_tools/planning_link_updater.h>
 #include <moveit/rviz_plugin_render_tools/robot_state_visualization.h>
